Initial bootstrap
This commit automates downloading and building nginx in an environment
contained within the project. It also adds start and stop scripts and an
very simple nginx.conf that gets symlinked in when the bootstrap script
is run. The bootstrap script can also be used to rebuild nginx if there
is any need to do so.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>