Adding GeoIP basics
This commit introduces a tiny rack application that returns the
inspected environment so that we can verify that the GeoIP module is
working properly. Nginx has also been configured to compile the geoip
module. The bootstrap script now downloads the GeoIP databases and
symlinks the extracted versions into the right place.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>